builder-stack-app-sample/requirements.txt
fastapi>=0.124.0
uvicorn>=0.38.0
gunicorn>=23.0.0
databricks-sdk>=0.73.0
pandas>=2.2.3
